Extract shared JSON export function
