+++ sql/moz-fx-data-shared-prod/telemetry_derived/newtab_merino_priors_to_gcs_v1/metadata.yaml
@@ -16,7 +16,7 @@ scheduling:
   - --source-dataset=telemetry_derived
   - --source-table=newtab_merino_priors_v1
   - --destination-bucket=merino-airflow-data-prodpy
-  - --destination-prefix=newtab-merino-priors
+  - --destination-prefix=newtab-merino-exports/priors
   - --deletion-days-old=3
 bigquery: null
 references: {}
